moved db setup out of the server.py module
